fix(web): make main library safe to import on browser/wasm targets
Previously, importing `package:dartastic_opentelemetry/dartastic_opentelemetry.dart`
from a Flutter web app or anything compiled via `dart compile js` was a
runtime trap. The main library re-exported the OTLP/HTTP exporters, the
certificate utilities, and the platform resource detectors — all of
which `import 'dart:io'` directly. `dart compile js` accepted the
imports thanks to Dart 3 stubs, but the moment any of those classes
ran (`HttpClient`, `SecurityContext`, `Platform.executable` etc.), you
got `UnsupportedError`.

Split the IO-bound code into platform-conditional facades, mirroring the
existing `gzip_io` / `gzip_web` / `gzip.dart` pattern:

- `lib/src/resource/native_detectors.dart` — exports `Process`/
  `HostResourceDetector` from `_io.dart` on native, `_stub.dart` on
  web. The stubs throw with a clear migration message if instantiated;
  in practice `PlatformResourceDetector.create()` only adds them on
  native, so the stubs are never reached.
- `lib/src/trace/export/otlp/certificate_utils.dart` — `_io.dart`
  keeps `validateCertificates` + `createSecurityContext`; `_stub.dart`
  keeps only `validateCertificates`. The IO-only `createSecurityContext`
  remains reachable from the IO HTTP exporter path. The gRPC exporters
  import `certificate_utils_io.dart` directly (gRPC over HTTP/2
  trailers isn't a thing in browsers regardless of dart:io).
- `lib/src/trace/export/otlp/http/http_client_factory.dart` — new
  helper. Native build returns `IOClient(HttpClient(context: ...))`
  with the configured `SecurityContext`; web build returns
  `BrowserClient` and logs a warning if the caller passed cert/key
  config (browser owns TLS — custom CA/mTLS isn't supported on the
  fetch path). The three `OtlpHttpSpanExporter`/`MetricExporter`/
  `LogRecordExporter` lost their direct `dart:io` imports and now
  delegate `_createHttpClient()` to this factory.

Net effect on web:
  - tracer/metrics/logs API works
  - OTLP/HTTP exporters work via browser fetch
  - `PlatformResourceDetector.create()` returns env-var + web detector
    composite (no IO detectors)
  - `OtlpGrpcSpanExporter` and friends remain native-only — fundamental
    to gRPC, not a fixable thing

Tests:
- New `test/web/web_compile_smoke_test.dart` runs in Chrome, imports
  the main library, initializes the SDK, constructs all three OTLP
  HTTP exporters, and runs the platform resource detector. Caught
  this regression class going forward.
- `test/unit/trace/export/otlp/certificate_utils_test.dart`,
  `test/unit/exporters_coverage_test.dart`,
  `test/unit/infrastructure_coverage_test.dart` — repointed at
  `certificate_utils_io.dart` directly (they exercise IO-only
  `createSecurityContext`); the certificate_utils test now declares
  `@TestOn('vm')`.

CHANGELOG entry added under `## [1.1.0-beta.1-wip]`.
