+++ lib/src/trace/export/otlp/http/otlp_http_span_exporter.dart
@@ -2,7 +2,6 @@
 // Copyright 2025, Michael Bushe, All rights reserved.
 
 import 'dart:async';
-import 'dart:io';
 import 'dart:math';
 import 'dart:typed_data';
 
@@ -10,13 +9,12 @@ import 'package:dartastic_opentelemetry/src/trace/span.dart';
 import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart'
     show OTelLog;
 import 'package:http/http.dart' as http;
-import 'package:http/io_client.dart';
 
 import '../../../../util/zip/gzip.dart';
 import '../../../span_logger.dart';
 import '../../span_exporter.dart';
-import '../certificate_utils.dart';
 import '../span_transformer.dart';
+import 'http_client_factory.dart';
 import 'otlp_http_span_exporter_config.dart';
 
 /// An OpenTelemetry span exporter that exports spans using OTLP over HTTP/protobuf
@@ -58,42 +56,15 @@ class OtlpHttpSpanExporter implements SpanExporter {
 
   /// Creates an HTTP client with custom certificates if configured.
   ///
-  /// This method creates an HttpClient with a SecurityContext configured
-  /// with any custom certificates specified in the exporter configuration.
-  http.Client _createHttpClient() {
-    // If no certificates are configured, use the default client
-    if (_config.certificate == null &&
-        _config.clientKey == null &&
-        _config.clientCertificate == null) {
-      return http.Client();
-    }
-
-    try {
-      final context = CertificateUtils.createSecurityContext(
+  /// Delegated to a platform-conditional factory: native gets an
+  /// `IOClient` wrapping an `HttpClient` with a custom `SecurityContext`;
+  /// web gets a `BrowserClient` (the browser handles TLS).
+  http.Client _createHttpClient() => createOtlpHttpClient(
+        exporterName: 'OtlpHttpSpanExporter',
         certificate: _config.certificate,
         clientKey: _config.clientKey,
         clientCertificate: _config.clientCertificate,
       );
-
-      if (context == null) {
-        return http.Client();
-      }
-
-      // Create an HttpClient with the custom SecurityContext
-      final httpClient = HttpClient(context: context);
-
-      // Wrap in IOClient for use with the http package
-      return IOClient(httpClient);
-    } catch (e) {
-      if (OTelLog.isError()) {
-        OTelLog.error(
-          'OtlpHttpSpanExporter: Failed to create HTTP client with certificates: $e',
-        );
-      }
-      // Fall back to default client on error
-      return http.Client();
-    }
-  }
 
   Duration _calculateJitteredDelay(int retries) {
     final baseMs = _config.baseDelay.inMilliseconds;
